split base and child
